9.2.0a3

+++ lib/lwres/getaddrinfo.c
@@ -19,18 +19,9 @@
  * PERFORMANCE OF THIS SOFTWARE.
  */
 
-/* $Id: getaddrinfo.c,v 1.38 2001/02/08 00:31:21 bwelling Exp $ */
+/* $Id: getaddrinfo.c,v 1.40 2001/07/10 18:25:40 gson Exp $ */
 
 #include <config.h>
-
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <sys/un.h>
-
-#include <netinet/in.h>
-
-#include <arpa/nameser.h>
-#include <arpa/inet.h>
 
 #include <string.h>
 #include <errno.h>
@@ -38,7 +29,7 @@
 
 #include <lwres/lwres.h>
 #include <lwres/net.h>
-#include <lwres/netdb.h>	/* XXX #include <netdb.h> */
+#include <lwres/netdb.h>
 
 #define SA(addr)	((struct sockaddr *)(addr))
 #define SIN(addr)	((struct sockaddr_in *)(addr))
@@ -181,7 +172,7 @@ lwres_getaddrinfo(const char *hostname, const char *servname,
 				return (EAI_SOCKTYPE);
 			if (port < 0 || port > 65535)
 				return (EAI_SERVICE);
-			port = htons(port);
+			port = htons((unsigned short) port);
 		} else {
 			sp = getservbyname(servname, proto);
 			if (sp == NULL)
